move normalizeMerge into defineFetch
