+++ src/dataServiceReducer/sharedReducer.test.js
@@ -1,4 +1,4 @@
-import sharedReducer, { normalizeMerge, replace } from './sharedReducer';
+import sharedReducer from './sharedReducer';
 import defineFetch from '../defineFetch';
 import { isSuccessAction } from '../createDataService';
 import SUCCESS from '../prefixes/SUCCESS';
@@ -7,172 +7,108 @@ import clearFetch from '../clearFetch';
 
 jest.mock('shortid', () => () => 'test-short-id');
 
-describe('normalizeMerge', () => {
-  const namespace = 'testNamespace';
+test('returns the previous state if the action is not a success action', () => {
+  // given
+  const otherAction = { type: 'test action' };
 
-  test('if falsy, returns replace', () => {
-    const result = normalizeMerge(null, namespace);
-    expect(result).toMatchInlineSnapshot(`
-          Object {
-            "testNamespace": [Function],
-          }
-        `);
-    expect(result[namespace]).toBe(replace);
-  });
+  const previousState = {};
 
-  test('if merge is a function, returns an object with the correct namespace', () => {
-    const result = normalizeMerge(() => ({}), namespace);
-    expect(result).toMatchInlineSnapshot(`
-          Object {
-            "testNamespace": [Function],
-          }
-        `);
-    expect(result[namespace]).not.toBe(replace);
-  });
+  // when
+  const newState = sharedReducer(previousState, otherAction);
 
-  test('if merge is an object and does not contain the current namespace, returns an object with replace', () => {
-    const result = normalizeMerge(
-      {
-        otherNamespace: () => ({}),
-      },
-      namespace,
-    );
-    expect(result).toMatchInlineSnapshot(`
-      Object {
-        "otherNamespace": [Function],
-        "testNamespace": [Function],
-      }
-      `);
-    expect(result[namespace]).toBe(replace);
-  });
-
-  test('if merge is an object but does contain the current namespace, returns the object as-is', () => {
-    const mergeFn = () => ({});
-    const mergeObj = {
-      [namespace]: mergeFn,
-    };
-    const result = normalizeMerge(mergeObj, namespace);
-
-    expect(result).toMatchInlineSnapshot(`
-      Object {
-        "testNamespace": [Function],
-      }
-      `);
-    expect(result).toBe(mergeObj);
-  });
-
-  test('if nothing matches, throw', () => {
-    expect(() => {
-      normalizeMerge('test', namespace);
-    }).toThrowErrorMatchingInlineSnapshot(
-      `"[sharedReducer] Could not match typeof merge. See docs. (TODO add docs link)"`,
-    );
-  });
+  // then
+  expect(newState).toBe(previousState);
 });
 
-describe('sharedReducer', () => {
-  test('returns the previous state if the action is not a success action', () => {
-    // given
-    const otherAction = { type: 'test action' };
-
-    const previousState = {};
-
-    // when
-    const newState = sharedReducer(previousState, otherAction);
-
-    // then
-    expect(newState).toBe(previousState);
+test('returns the previous state if the action is not shared', () => {
+  // given
+  const actionCreatorFactory = defineFetch({
+    displayName: 'test action',
+    make: testArg => ({
+      key: [testArg],
+      request: () => () => {},
+    }),
   });
 
-  test('returns the previous state if the action is not shared', () => {
-    // given
-    const actionCreatorFactory = defineFetch({
-      displayName: 'test action',
-      make: testArg => ({
-        key: [testArg],
-        request: () => () => {},
-      }),
-    });
+  const fetchAction = actionCreatorFactory()();
+  const successAction = {
+    type: createActionType(SUCCESS, fetchAction.meta),
+    meta: fetchAction.meta,
+    payload: { mock: 'data' },
+  };
 
-    const fetchAction = actionCreatorFactory()();
-    const successAction = {
-      type: createActionType(SUCCESS, fetchAction.meta),
-      meta: fetchAction.meta,
-      payload: { mock: 'data' },
-    };
+  expect(isSuccessAction(successAction)).toBe(true);
 
-    expect(isSuccessAction(successAction)).toBe(true);
+  const previousState = {};
 
-    const previousState = {};
+  // when
+  const newState = sharedReducer(previousState, successAction);
 
-    // when
-    const newState = sharedReducer(previousState, successAction);
+  // then
+  expect(newState).toBe(previousState);
+});
 
-    // then
-    expect(newState).toBe(previousState);
+test.only('does merges correctly', () => {
+  const makeMovieListFetch = defineFetch({
+    displayName: 'Get Movie List',
+    share: {
+      namespace: 'movieList',
+      merge: {
+        movieItem: (previousList, movie) => {
+          if (!previousList) return null;
+
+          const index = previousList.findIndex(i => i.id === movie.id);
+          if (index === -1) return previousList;
+
+          return [
+            ...previousList.slice(0, index),
+            movie,
+            ...previousList.slice(index + 1, previousList.length),
+          ];
+        },
+      },
+    },
+    make: () => ({
+      key: [],
+      request: () => ({ http }) =>
+        http({
+          method: 'GET',
+          route: '/movies',
+        }),
+    }),
+  });
+  const makeMovieItemFetch = defineFetch({
+    displayName: 'Get Movie Item',
+    share: {
+      namespace: 'movieItem',
+      merge: {
+        // movieList: (previousMovie, nextList) => {
+        //   if (!previousMovie) return null;
+        //   return nextList.find(i => i.id === previousMovie.id);
+        // },
+      },
+    },
+    make: movieId => ({
+      key: [movieId],
+      request: () => ({ http }) =>
+        http({
+          method: 'GET',
+          route: `/movies/${movieId}`,
+        }),
+    }),
   });
 
-  test.only('does merges correctly', () => {
-    const makeMovieListFetch = defineFetch({
-      displayName: 'Get Movie List',
-      share: {
-        namespace: 'movieList',
-        merge: {
-          movieItem: (previousList, movie) => {
-            if (!previousList) return null;
+  const movieListFetch = makeMovieListFetch();
+  const movieItem123Fetch = makeMovieItemFetch('movie123');
 
-            const index = previousList.findIndex(i => i.id === movie.id);
-            if (index === -1) return previousList;
+  const movieListRequest = movieListFetch();
+  const movieItem123Request = movieItem123Fetch();
 
-            return [
-              ...previousList.slice(0, index),
-              movie,
-              ...previousList.slice(index + 1, previousList.length),
-            ];
-          },
-        },
-      },
-      make: () => ({
-        key: [],
-        request: () => ({ http }) =>
-          http({
-            method: 'GET',
-            route: '/movies',
-          }),
-      }),
-    });
-    const makeMovieItemFetch = defineFetch({
-      displayName: 'Get Movie Item',
-      share: {
-        namespace: 'movieItem',
-        merge: {
-          // movieList: (previousMovie, nextList) => {
-          //   if (!previousMovie) return null;
-          //   return nextList.find(i => i.id === previousMovie.id);
-          // },
-        },
-      },
-      make: movieId => ({
-        key: [movieId],
-        request: () => ({ http }) =>
-          http({
-            method: 'GET',
-            route: `/movies/${movieId}`,
-          }),
-      }),
-    });
+  const requests = [movieListRequest, movieItem123Request];
 
-    const movieListFetch = makeMovieListFetch();
-    const movieItem123Fetch = makeMovieItemFetch('movie123');
+  const afterRequests = requests.reduce(sharedReducer, null);
 
-    const movieListRequest = movieListFetch();
-    const movieItem123Request = movieItem123Fetch();
-
-    const requests = [movieListRequest, movieItem123Request];
-
-    const afterRequests = requests.reduce(sharedReducer, null);
-
-    expect(afterRequests).toMatchInlineSnapshot(`
+  expect(afterRequests).toMatchInlineSnapshot(`
       Object {
         "merges": Object {
           "movieItem": Object {
@@ -202,17 +138,17 @@ describe('sharedReducer', () => {
       }
     `);
 
-    const movie123Success = {
-      type: createActionType(SUCCESS, movieItem123Fetch.meta),
-      meta: movieItem123Fetch.meta,
-      payload: {
-        id: 'movie123',
-        title: 'foo',
-      },
-    };
+  const movie123Success = {
+    type: createActionType(SUCCESS, movieItem123Fetch.meta),
+    meta: movieItem123Fetch.meta,
+    payload: {
+      id: 'movie123',
+      title: 'foo',
+    },
+  };
 
-    const afterMovie123Success = sharedReducer(afterRequests, movie123Success);
-    expect(afterMovie123Success).toMatchInlineSnapshot(`
+  const afterMovie123Success = sharedReducer(afterRequests, movie123Success);
+  expect(afterMovie123Success).toMatchInlineSnapshot(`
       Object {
         "data": Object {
           "movieItem": Object {
@@ -251,23 +187,23 @@ describe('sharedReducer', () => {
       }
     `);
 
-    const movieListSuccess = {
-      type: createActionType(SUCCESS, movieListFetch.meta),
-      meta: movieListFetch.meta,
-      payload: [
-        {
-          id: 'movie123',
-          title: 'CHANGED',
-        },
-        {
-          id: 'movie456',
-          title: 'foo',
-        },
-      ],
-    };
+  const movieListSuccess = {
+    type: createActionType(SUCCESS, movieListFetch.meta),
+    meta: movieListFetch.meta,
+    payload: [
+      {
+        id: 'movie123',
+        title: 'CHANGED',
+      },
+      {
+        id: 'movie456',
+        title: 'foo',
+      },
+    ],
+  };
 
-    const afterMoveListSuccess = sharedReducer(afterMovie123Success, movieListSuccess);
-    expect(afterMoveListSuccess).toMatchInlineSnapshot(`
+  const afterMoveListSuccess = sharedReducer(afterMovie123Success, movieListSuccess);
+  expect(afterMoveListSuccess).toMatchInlineSnapshot(`
       Object {
         "data": Object {
           "movieItem": Object {
@@ -316,30 +252,30 @@ describe('sharedReducer', () => {
         },
       }
     `);
+});
+
+test.skip('clear fetch', () => {
+  // given
+  const fetchFactory = defineFetch({
+    displayName: 'Get People',
+    share: { namespace: 'people' },
+    make: personId => ({
+      key: [personId],
+      request: () => () => ({ id: personId, foo: 'bar' }),
+    }),
   });
 
-  test.skip('clear fetch', () => {
-    // given
-    const fetchFactory = defineFetch({
-      displayName: 'Get People',
-      share: { namespace: 'people' },
-      make: personId => ({
-        key: [personId],
-        request: () => () => ({ id: personId, foo: 'bar' }),
-      }),
-    });
+  const fetch = fetchFactory('person123');
+  const fetchAction = fetch();
 
-    const fetch = fetchFactory('person123');
-    const fetchAction = fetch();
+  const successAction = {
+    type: createActionType(SUCCESS, fetchAction.meta),
+    meta: fetchAction.meta,
+    payload: { mock: 'data' },
+  };
 
-    const successAction = {
-      type: createActionType(SUCCESS, fetchAction.meta),
-      meta: fetchAction.meta,
-      payload: { mock: 'data' },
-    };
-
-    const state = sharedReducer({}, successAction);
-    expect(state).toMatchInlineSnapshot(`
+  const state = sharedReducer({}, successAction);
+  expect(state).toMatchInlineSnapshot(`
       Object {
         "people | key:person123": Object {
           "data": Object {
@@ -349,12 +285,11 @@ describe('sharedReducer', () => {
       }
     `);
 
-    const clearAction = clearFetch(fetch);
+  const clearAction = clearFetch(fetch);
 
-    // when
-    const newState = sharedReducer(state, clearAction);
+  // when
+  const newState = sharedReducer(state, clearAction);
 
-    // then
-    expect(newState).toMatchInlineSnapshot(`Object {}`);
-  });
+  // then
+  expect(newState).toMatchInlineSnapshot(`Object {}`);
 });
